+++ NEWS.md
@@ -1,6 +1,7 @@
 ## CHANGES IN doc2vec VERSION 0.2.1
 
 - Make sure words are only 100 characters when getting embeddings of documents (issue #20)
+- Limit documents to 1000 words by explicitely keeping only the first 1000 words per document + provide warning if doc_id contains spaces
 
 ## CHANGES IN doc2vec VERSION 0.2.0